+++ src/components/admin-login/admin-login.component.tsx
@@ -1,4 +1,4 @@
-import HankoAuth from "../hanko-auth/hanko-auth.component";
+import HankoAuth from "../../utilities/hanko/hanko-auth.component";
 
 const AdminLogin = () => {
   return <HankoAuth />
